updated home and dashboard pages, added api key security
