+++ client/package.json
@@ -17,7 +17,8 @@
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.4",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.11.2"
+    "react-router-dom": "^6.11.2",
+    "dotenv": "^8.2.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.28",
