+++ client/src/App.jsx
@@ -26,11 +26,6 @@ const authLink = setContext((_, { headers }) => {
 
 
 
-//const client = new ApolloClient({
-//  uri: '/graphql',
-//  cache: new InMemoryCache(),
-//});
-
 const client = new ApolloClient({
   // Set up our client to execute the `authLink` middleware prior to making the request to our GraphQL API
   link: authLink.concat(httpLink),
