+++ client/src/utils/queries.js
@@ -19,4 +19,31 @@ export const GET_ME = gql`
       }
     }
   }
-`;
+`
+export const QUERY_CITIES = gql`
+  query getCities {
+    cities{
+      _id
+      username
+      email
+      cityCount
+      savedCities {
+        cityId
+        formattedAddress
+        cityName
+        countyName
+        stateName
+        countryName
+        latitude
+        longitude 
+      }
+    }
+  }`
+
+  export const QUERY_APIKEY = gql`
+  query getAPIKey {
+    apikey{
+      apikey
+    }
+  }`
+;
