+++ server/.env
@@ -1,0 +1,1 @@
+GOOGLE_API_KEY="AIzaSyDwkYPjXsxzUOY4rsDbJh0yY_uksIYMPY0"
